Ignore diacritics in FinalDropdown search and apply on CountryInput
